requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.8
pymongo==3.11.0
python-dateutil==2.8.1
python-dotenv==0.13.0
